Fix: repetitive notes (with pedal) not always audible
If the voice limit's exceeded only the newest X notes play. However, when a note is hit again (meaning it's not an all new note but a "refresh" of an existing one) it wouldn't move to the top of the list and thus not be part of the X notes that are actually audible.