Support for "headless" operation and passthrough-mode fix
Syntherrupter would only enter passthrough mode if *no* Nextion screen at all is connected. Which doesn't make too much sense. On the other hand, Syntherrupter would *not* enter passthrough mode if a screen is connected but has no or the wrong firmware.
Now, when no screen is connected, syntherrupter can be controlled via sysex. And if a screen is connected, it works as one would expect.