Fixed: Race condition causes corrupted notelist (hanging tones)
On the second byte of an note on command the associated note object is retrieved - if there is any. However, because of the per-byte processing of MIDI commands, the function returns. If the next byte is not yet available, other things are processed. If on top of that the specific note retrieved above ends right now it gets removed from the notelist.
Now, on the third byte of the note on command, the note is marked and treated as existing even though it's not, causing the notelist to become corrupted. Hanging tones are the result in this case.
The fix is simple: make the Note On Event handling atomic. Do both retrieving and modifying without returning inbetween.